+++ src/components/Store/Store.tsx
@@ -52,18 +52,18 @@ export const Store: React.FC<{
   }, []);
 
   const handleNewProject = useCallback(() => {
-    const newBoardId = adapter.createDocument();
+    const newBoardId = adapter.document.create();
     navigate(`/board/${newBoardId}`);
   }, [adapter, navigate]);
 
   const handleOpenProject = useCallback(async () => {
     const fileContents = await fileSystem.openFile();
-    const newBoardId = adapter.loadDocument(fileContents);
+    const newBoardId = adapter.document.load(fileContents);
     navigate(`/board/${newBoardId}`);
   }, [adapter]);
 
   const handleSaveProject = useCallback(async () => {
-    const fileContents = adapter.serialiseDocument();
+    const fileContents = adapter.document.serialise();
     await fileSystem.saveFile(fileContents);
   }, [adapter]);